# docs/openapi.yaml
openapi: 3.0.3
info:
  title: АИС «Реестр обязательных требований» — HTTP API
  version: "1.0.0"
  description: |
    Спецификация фактического API веб-приложения (Next.js App Router, 46 маршрутов).
    Составлена по коду `src/app/api/**` (2026-07-02), см. docs/architecture (комплект ГОСТ 34).

    **Аутентификация:** JWT в httpOnly-cookie `auth_token` (выдаётся `/api/auth/login`, срок 7 суток).
    **Роли:** `admin` (супер-админ МНЭ) · `mne` (сотрудник МНЭ: всё как admin, кроме выгрузки каталога и создания учёток уровня МНЭ) · `moderator` (орган + поддерево организаций) ·
    `expert` (назначенные сферы ∩ органы). Скоупы возвращает `/api/auth/me`.
    **Ошибки:** единый формат `{ "error": string }`; 401 — не авторизован, 403 — нет прав/чужой орган,
    404 — не найдено, 429 — превышен лимит (голосование), 5xx — внутренняя/внешний сервис.

    Раздел «Легаси-экспертиза» обслуживает исторический MVP-контур (requirement_cards/голосования)
    и не развивается — см. документ 09 плана (К5).
servers:
  - url: https://reestr-web-production.up.railway.app
    description: Продуктив (Railway)
  - url: http://localhost:3000
    description: Локальная разработка
security:
  - cookieAuth: []
...
paths:
...
  /api/npa-submission/preview:
    post:
      tags: [Подача НПА]
      summary: Мгновенное ИИ-превью НПА (~15 с, без записи в реестр)
      description: |
        Node загружает adilet-HTML по ngr → чанкер (src/lib/npaParse.ts, каскад «ст.N»/«п.zN»/окна) →
        DeepSeek по первым содержательным сегментам → черновые требования.
        Требует DEEPSEEK_API_KEY на сервере.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ngr]
              properties: { ngr: { type: string } }
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  title: { type: string }
                  articleCount: { type: integer }
                  previewedArticles: { type: array, items: { type: string }, description: "Метки сегментов («ст.5», «п.z16»)" }
                  requirements:
                    type: array
                    items:
                      type: object
                      properties:
                        subject: { type: string }
                        action: { type: string }
                        object: { type: string, nullable: true }
                        condition: { type: string, nullable: true }
                        quote: { type: string }
                  note: { type: string, description: Если структура не распознана }
        "502": { description: Сбой adilet/DeepSeek }

  # Бизнес-портал вынесен в отдельный публичный сервис reestr-business
  # (репо YergZakon/reestr-business, https://reestr-business-production.up.railway.app).
  /api/npa-assignment:
    get:
      tags: [Ревью и АРА]
      summary: НПА органа с текущим назначением комитета (+комитеты, +журнал)
      parameters:
        - { name: org_id, in: query, required: true, schema: { type: integer }, description: Узел министерства }
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string, enum: [all, assigned, unassigned] } }
        - { name: page, in: query, schema: { type: integer } }
      responses:
        "200": { description: "items[], committees[], log[]" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Ревью и АРА]
      summary: Назначить ответственный комитет за НПА (каскад на все требования)
      description: Закрывает прежнее активное назначение; authority_code всех требований НПА переводится на код комитета; комитет получает уведомление.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ngr, org_id]
              properties:
                ngr: { type: string }
                org_id: { type: integer, description: Узел комитета }
                reason: { type: string, maxLength: 500, description: Основание назначения }
      responses:
        "200": { description: "{ ok, assignment_id, cascaded, committee }" }
        "403": { $ref: "#/components/responses/Forbidden" }
    delete:
      tags: [Ревью и АРА]
      summary: Отменить активное назначение (требования возвращаются министерству)
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [ngr], properties: { ngr: { type: string } } }
      responses:
        "200": { description: "{ ok, cascaded, back_to }" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT (7 суток), выдаётся /api/auth/login
  responses:
...
    Forbidden:
      description: Недостаточно прав / чужой орган
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Не найдено
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...